Remove pending repo and merge developer packages
