+++ repos/exawind/packages/amr-wind/package.py
@@ -1,12 +1,9 @@
 from spack import *
-from spack.pkg.pending.amr_wind import AmrWind as bAmrWind
+from spack.pkg.builtin.amr_wind import AmrWind as bAmrWind
 import os
 from shutil import copyfile
 
-class AmrWindDeveloper(bAmrWind):
-    variant('compile_commands', default=True,
-            description='generate compile_commands.json and copy to source dir')
-
+class AmrWind(bAmrWind):
     depends_on('ninja', type='build')
     generator = 'Ninja'
 
@@ -18,7 +15,7 @@ class AmrWindDeveloper(bAmrWind):
         if spec['mpi'].name == 'openmpi':
             options.append(define('MPIEXEC_PREFLAGS', '--oversubscribe'))
 
-        if '+compile_commands' in spec:
+        if 'dev_path' in spec:
             options.append(define('CMAKE_EXPORT_COMPILE_COMMANDS',True))
 
         saved_golds = os.path.join(os.getenv('SPACK_MANAGER'), 'tmp', 'tmp_golds', 'amr-wind')
@@ -33,7 +30,7 @@ class AmrWindDeveloper(bAmrWind):
 
     @run_after('cmake')
     def copy_compile_commands(self):
-        if '+compile_commands' in self.spec:
+        if 'dev_path' in self.spec:
             target = os.path.join(self.stage.source_path, "compile_commands.json")
             source = os.path.join(self.build_directory, "compile_commands.json")
             copyfile(source, target)
